+++ pyodide_build/pywasmcross.py
@@ -193,6 +193,8 @@ def replay_genargs_handle_linker_opts(arg: str) -> str | None:
         # macOS-specific linker flags that wasm-ld doesn't understand
         "-headerpad_max_install_names",
         "-dead_strip_dylibs",
+        # Newer clang seems to add this by default, but wasm-ld doesn't support it
+        "--enable-new-dtags",
     )
 
     excluded_linker_prefixes = (
@@ -560,7 +562,13 @@ def handle_command_generate_args(  # noqa: C901
         line[0] = "emranlib"
         return line
     elif cmd == "strip":
-        line[0] = "emstrip"
+        if build_args.abi > "2026":
+            # Emscripten > 5.0.0 strips out the dylink.0 section
+            # so we skip strip to avoid the error
+            # https://github.com/emscripten-core/emscripten/issues/26563
+            line[0] = "echo"
+        else:
+            line[0] = "emstrip"
         return line
     else:
         return line